+++ packages/web-vue/components/image/preview-group.vue
@@ -11,7 +11,11 @@
     :popup-container="popupContainer"
     :render-to-body="renderToBody"
     @close="onClose"
-  />
+  >
+    <template v-if="$slots.actions" #actions>
+      <slot name="actions" :url="currentUrl" />
+    </template>
+  </ImagePreview>
 </template>
 <script lang="tsx">
 import {
@@ -142,6 +146,12 @@ export default defineComponent({
     'visible-change',
     'update:visible',
   ],
+  /**
+   * @zh 自定义额外的操作项
+   * @en Customize additional action items
+   * @slot actions
+   * @version 2.46.0
+   */
   setup(props: ImagePreviewGroupProps, { emit }) {
     const {
       srcList,