+++ packages/web-vue/components/select/select-dropdown.vue
@@ -1,13 +1,16 @@
 <template>
   <div :class="cls">
+    <div
+      v-if="$slots.header && (!empty || showHeaderOnEmpty)"
+      :class="`${prefixCls}-header`"
+    >
+      <slot name="header" />
+    </div>
     <spin v-if="loading" :class="`${prefixCls}-loading`" />
     <div v-else-if="empty" :class="`${prefixCls}-empty`">
       <slot name="empty">
-        <empty />
+        <component :is="SelectEmpty ? SelectEmpty : 'Empty'" />
       </slot>
-    </div>
-    <div v-if="$slots.header && !empty" :class="`${prefixCls}-header`">
-      <slot name="header" />
     </div>
     <slot v-if="virtualList && !loading && !empty" name="virtual-list" />
     <component
@@ -23,7 +26,10 @@
         <slot />
       </ul>
     </component>
-    <div v-if="$slots.footer && !empty" :class="`${prefixCls}-footer`">
+    <div
+      v-if="$slots.footer && (!empty || showFooterOnEmpty)"
+      :class="`${prefixCls}-footer`"
+    >
       <slot name="footer" />
     </div>
   </div>
@@ -31,8 +37,9 @@
 
 <script lang="ts">
 import type { PropType } from 'vue';
-import { computed, defineComponent, toRefs } from 'vue';
+import { computed, defineComponent, toRefs, inject } from 'vue';
 import type { EmitType } from '../_utils/types';
+import { configProviderInjectionKey } from '../config-provider/context';
 import { getPrefixCls } from '../_utils/global-config';
 import Empty from '../empty';
 import Spin from '../spin';
@@ -65,11 +72,22 @@ export default defineComponent({
     onReachBottom: {
       type: [Function, Array] as PropType<EmitType<(ev: Event) => void>>,
     },
+    showHeaderOnEmpty: {
+      type: Boolean as PropType<boolean>,
+      default: false,
+    },
+    showFooterOnEmpty: {
+      type: Boolean as PropType<boolean>,
+      default: false,
+    },
   },
   emits: ['scroll', 'reachBottom'],
   setup(props, { emit, slots }) {
     const { scrollbar } = toRefs(props);
     const prefixCls = getPrefixCls('select-dropdown');
+    const configCtx = inject(configProviderInjectionKey, undefined);
+    const SelectEmpty = configCtx?.slots.empty?.({ component: 'select' })?.[0];
+
     const { componentRef: wrapperComRef, elementRef: wrapperRef } =
       useComponentRef('containerRef');
     const { displayScrollbar, scrollbarProps } = useScrollbar(scrollbar);
@@ -93,6 +111,7 @@ export default defineComponent({
 
     return {
       prefixCls,
+      SelectEmpty,
       cls,
       wrapperRef,
       wrapperComRef,