+++ packages/web-vue/components/tree-select/tree-select.vue
@@ -47,14 +47,24 @@
     </slot>
     <template #content>
       <div
-        :class="[`${prefixCls}-popup`, dropdownClassName]"
+        :class="[
+          `${prefixCls}-popup`,
+          {
+            [`${prefixCls}-has-header`]: Boolean($slots.header),
+            [`${prefixCls}-has-footer`]: Boolean($slots.footer),
+          },
+          dropdownClassName,
+        ]"
         :style="computedDropdownStyle"
       >
+        <div v-if="$slots.header && !isEmpty" :class="`${prefixCls}-header`">
+          <slot name="header" />
+        </div>
         <slot v-if="loading" name="loader">
           <Spin />
         </slot>
-        <slot v-else-if="isEmptyTreeData || isEmptyFilterResult" name="empty">
-          <Empty />
+        <slot v-else-if="isEmpty" name="empty">
+          <component :is="TreeSelectEmpty ? TreeSelectEmpty : 'Empty'" />
         </slot>
         <Panel
           v-else
@@ -79,6 +89,9 @@
           :tree-slots="pickSubCompSlots($slots, 'tree')"
           @change="onSelectChange"
         />
+        <div v-if="$slots.footer && !isEmpty" :class="`${prefixCls}-footer`">
+          <slot name="footer" />
+        </div>
       </div>
     </template>
   </Trigger>
@@ -94,6 +107,7 @@ import {
   ref,
   toRefs,
   StyleValue,
+  inject,
 } from 'vue';
 import useMergeState from '../_hooks/use-merge-state';
 import { LabelValue } from './interface';
@@ -101,6 +115,7 @@ import Trigger, { TriggerProps } from '../trigger';
 import SelectView from '../_components/select-view/select-view';
 import Panel from './panel';
 import { getPrefixCls } from '../_utils/global-config';
+import { configProviderInjectionKey } from '../config-provider/context';
 import useSelectedState from './hooks/use-selected-state';
 import useTreeData from '../tree/hooks/use-tree-data';
 import {
@@ -124,6 +139,7 @@ import {
 import { isNodeSelectable } from '../tree/utils';
 import { Data } from '../_utils/types';
 import { ScrollbarProps } from '../scrollbar';
+import { SelectViewValue } from '../_components/select-view/interface';
 
 export default defineComponent({
   name: 'TreeSelect',
@@ -478,7 +494,17 @@ export default defineComponent({
    * @en Render additional node content of the tree component
    * @slot tree-slot-extra
    */
-  setup(props, { emit }) {
+  /**
+   * @zh 自定义下拉框页头
+   * @en The header of the drop-down box
+   * @slot header
+   */
+  /**
+   * @zh 自定义下拉框页脚
+   * @en The footer of the drop-down box
+   * @slot footer
+   */
+  setup(props, { emit, slots }) {
     const {
       defaultValue,
       modelValue,
@@ -497,11 +523,16 @@ export default defineComponent({
       treeProps,
       fallbackOption,
       selectable,
+      dropdownClassName,
     } = toRefs(props);
     const { mergedDisabled, eventHandlers } = useFormItem({
       disabled,
     });
     const prefixCls = getPrefixCls('tree-select');
+    const configCtx = inject(configProviderInjectionKey, undefined);
+    const TreeSelectEmpty = configCtx?.slots.empty?.({
+      component: 'tree-select',
+    })?.[0];
     const isMultiple = computed(() => multiple.value || treeCheckable.value);
     const isSelectable = (
       node: TreeNodeData,
@@ -564,9 +595,9 @@ export default defineComponent({
             ...i,
             closable: !node || isNodeClosable(node),
           };
-        });
+        }) as SelectViewValue[];
       }
-      return selectedValue.value;
+      return selectedValue.value as SelectViewValue[];
     });
 
     const setSelectedKeys = (newVal: TreeNodeKey[]) => {
@@ -619,7 +650,9 @@ export default defineComponent({
         })
       );
 
-    const isEmptyTreeData = computed(() => !flattenTreeData.value.length);
+    const isEmpty = computed(
+      () => !flattenTreeData.value.length || isEmptyFilterResult.value
+    );
 
     const refSelectView = ref();
 
@@ -637,13 +670,13 @@ export default defineComponent({
     return {
       refSelectView,
       prefixCls,
+      TreeSelectEmpty,
       selectedValue,
       selectedKeys,
       mergedDisabled,
       searchValue,
       panelVisible,
-      isEmptyTreeData,
-      isEmptyFilterResult,
+      isEmpty,
       computedFilterTreeNode,
       isMultiple,
       selectViewValue,