+++ packages/web-vue/components/trigger/trigger.tsx
@@ -321,6 +321,15 @@ export default defineComponent({
       type: Boolean,
       default: false,
     },
+    /**
+     * @zh 是否在滚动时关闭弹出框
+     * @en Whether to close the popover when scrolling
+     * @version 2.46.0
+     */
+    scrollToClose: {
+      type: Boolean,
+      default: false,
+    },
   },
   emits: {
     'update:popupVisible': (visible: boolean) => true,
@@ -392,6 +401,7 @@ export default defineComponent({
 
     let delayTimer = 0;
     let outsideListener = false;
+    let windowListener = false;
 
     const cleanDelayTimer = () => {
       if (delayTimer) {
@@ -628,8 +638,22 @@ export default defineComponent({
 
     const handleScroll = throttleByRaf(() => {
       if (computedVisible.value) {
-        updatePopupStyle();
+        if (props.scrollToClose || configCtx?.scrollToClose) {
+          changeVisible(false);
+        } else {
+          updatePopupStyle();
+        }
       }
+    });
+
+    const removeWindowScroll = () => {
+      off(window, 'scroll', onWindowScroll);
+      windowListener = false;
+    };
+
+    const onWindowScroll = throttleByRaf(() => {
+      changeVisible(false);
+      removeWindowScroll();
     });
 
     const handleResize = () => {
@@ -666,6 +690,11 @@ export default defineComponent({
           on(document.documentElement, 'mousedown', handleOutsideClick);
           outsideListener = true;
         }
+      }
+
+      if (props.scrollToClose || configCtx?.scrollToClose) {
+        on(window, 'scroll', onWindowScroll);
+        windowListener = true;
       }
 
       if (props.updateAtScroll || configCtx?.updateAtScroll) {
@@ -736,6 +765,9 @@ export default defineComponent({
       destroyResizeObserver();
       if (outsideListener) {
         removeOutsideListener();
+      }
+      if (windowListener) {
+        removeWindowScroll();
       }
       if (scrollElements) {
         for (const item of scrollElements) {